Changes to packet retrieval and refactoring
The position of the current byte is no longer passed as a parameter when getting new packet data.

The 'main.py' file has been renamed to 'unpacker.py' and its sole purpose is now to connect to the socket and unpack received data. To use the data, other python scripts should import the unpacker file and use the RetrievePacket() method. The packet ID will still have to be checked to ensure data is from the expected packet type.

A test program, 'paceGraph.py' has been created to test the new system.