Added speed trace graph
A graph showing the speed to lap distance of each driver is shown on the window.

Each speed trace is coloured using their teams colour code (Current F1 2021 teams only, other teams may result in error). Drivers in the same team are differentiated with a different thickness of trace line.