Lookup data and fixed final classification packet error
Added a data lookup file that will contain constant data to be displayed in the app. Moved team colour data to here.

App now displays the session type, track name and location as well as the team name of the current driver.

Fixed an error caused by an incorrect buffer length when receiving the final classification packet.